customizacao da pagina de erro e revisao dos exceptions no controller da startup - Tocha

+++ protected/controllers/StartupController.php
@@ -37,12 +37,12 @@ class StartupController extends Controller
 				'users'=>array('@'),
 			),
             array('allow', // allow authenticated user to perform 'create' and 'update' actions
-				'actions'=>array('update', 'updateName', 'updateLocation', 'updateSectors', 'follow', 'unfollow', 'approve', 'addTeam', 'deleteTeam', 'RefreshStartupName'),
+				'actions'=>array('update', 'updateName', 'updateLocation', 'updateSectors', 'follow', 'unfollow', 'approve', 'addTeam', 'deleteTeam', 'RefreshStartupName', 'delete'),
 				'users'=>array('@'),
                 'verbs'=>array('POST'),
 			),
 			array('allow', // allow admin user to perform 'admin' and 'delete' actions
-				'actions'=>array('admin','delete'),
+				'actions'=>array('admin'),
 				'users'=>array('admin'),
 			),
 			array('deny',  // deny all users
@@ -69,7 +69,7 @@ class StartupController extends Controller
 		else
 		{
 			 if(!Yii::app()->user->checkAccess('editStartup', array('startup'=>$model)))
-				throw new CHttpException(404, 'Page not found.');
+				throw new CHttpException(404,UserModule::t('Page not found.'));
 			
 			else
 				$this->redirect(array('/edit/'.$model->startupname));
@@ -85,7 +85,7 @@ class StartupController extends Controller
 		$model->sec=$model->sectors;
 		
         if(!Yii::app()->user->checkAccess('editStartup', array('startup'=>$model)))
-            throw new CHttpException(403,'Você não pode editar essa startup!');
+            throw new CHttpException(403,UserModule::t('You cannot edit this Startup!'));
 		
 		if(!$model->sectors || !$model->product_description || !$model->company_stage)
 		{
@@ -136,7 +136,7 @@ class StartupController extends Controller
         $model=$this->loadModel($name);
        
         if(!Yii::app()->user->checkAccess('editStartup', array('startup'=>$model)))
-            throw new CHttpException(403,'Você não pode editar essa startup!');
+            throw new CHttpException(403,UserModule::t('You cannot edit this Startup!'));
 		
 		if(isset($_POST['Startup']['pic']))
 		{
@@ -430,7 +430,7 @@ class StartupController extends Controller
 	{
         $model=$this->loadModelId($_POST['pk']);
         if(!Yii::app()->user->checkAccess('editStartup', array('startup'=>$model)))
-            throw new CHttpException(403,'Você não pode editar esse projeto!');
+            throw new CHttpException(403,UserModule::t('You cannot edit this Startup!'));
 			
 		$es = new TbEditableSaver('Startup');  //'Startup' is name of model to be updated
         $es->update();
@@ -459,7 +459,7 @@ class StartupController extends Controller
 	{
         $model=$this->loadModelId($_POST['pk']);
         if(!Yii::app()->user->checkAccess('editStartup', array('startup'=>$model)))
-            throw new CHttpException(403,'Você não pode editar esse projeto!');
+            throw new CHttpException(403,UserModule::t('You cannot edit this Startup!'));
 			
 		$es = new TbEditableSaver('Startup');  //'Startup' is name of model to be updated
         $es->update();
@@ -495,7 +495,7 @@ class StartupController extends Controller
     {
         $model = $this->loadModelId($_POST['pk']);
 		if(!Yii::app()->user->checkAccess('editStartup', array('startup'=>$model)))
-            throw new CHttpException(403,'Você não pode editar esse projeto!');
+            throw new CHttpException(403,UserModule::t('You cannot edit this Startup!'));
 			
         if($_POST['value']==0) $model->location=NULL;
         else $model->location = $_POST['value'];
@@ -507,7 +507,7 @@ class StartupController extends Controller
     {
         $model = $this->loadModelId($_POST['pk']);
 		if(!Yii::app()->user->checkAccess('editStartup', array('startup'=>$model)))
-            throw new CHttpException(403,'Você não pode editar esse projeto!');
+            throw new CHttpException(403,UserModule::t('You cannot edit this Startup!'));
 			
         $vals = array();
         if(isset($_POST['value']))
@@ -537,17 +537,26 @@ class StartupController extends Controller
 	*/
 	public function actionDelete($id)
 	{
+		$model = $this->loadModelId($id);
+		if(!Yii::app()->user->checkAccess('editStartup', array('startup'=>$model)))
+            throw new CHttpException(403,UserModule::t('You cannot edit this Startup!'));
+		
 		if(Yii::app()->request->isPostRequest)
 		{
 			// we only allow deletion via POST request
 			$this->loadModelId($id)->delete();
 
-			// if AJAX request (triggered by deletion via admin grid view), we should not redirect the browser
-			if(!isset($_GET['ajax']))
-				$this->redirect(isset($_POST['returnUrl']) ? $_POST['returnUrl'] : array('admin'));
+			if(Yii::app()->user->isSuperuser)
+			{
+				// if AJAX request (triggered by deletion via admin grid view), we should not redirect the browser
+				if(!isset($_GET['ajax']))
+					$this->redirect(isset($_POST['returnUrl']) ? $_POST['returnUrl'] : array('admin'));
+			}
+			else
+				$this->redirect(Yii::app()->request->baseUrl.'/'.Yii::app()->user->username);
 		}
 		else
-			throw new CHttpException(400,'Invalid request. Please do not repeat this request again.');
+			throw new CHttpException(404,UserModule::t('It was not possible to resolve the request.'));
 	}
 
 	/**
@@ -640,7 +649,7 @@ class StartupController extends Controller
 										  ':name'=>$name,
 										));
 		if($model===null)
-			throw new CHttpException(404,'The requested page does not exist.');
+			throw new CHttpException(404,UserModule::t('Page not found.'));
 		return $model;
 	}
 	
@@ -651,7 +660,7 @@ class StartupController extends Controller
 										  ':id'=>$id,
 										));
 		if($model===null)
-			throw new CHttpException(404,'The requested page does not exist.');
+			throw new CHttpException(404,UserModule::t('Page not found.'));
 		return $model;
 	}
 	
@@ -673,7 +682,7 @@ class StartupController extends Controller
     {	
 		$model = $this->loadModel($_GET['name']);
 		if(!Yii::app()->user->checkAccess('editStartup', array('startup'=>$model)))
-            throw new CHttpException(403,'Você não pode editar esse projeto!');
+            throw new CHttpException(403,UserModule::t('You cannot edit this Startup!'));
          
         if(isset($_POST['sectors']))
         {                    
@@ -731,7 +740,7 @@ class StartupController extends Controller
 		}
 		
 		else
-			throw new CHttpException(404, 'Page not found.');
+			throw new CHttpException(404,UserModule::t('Page not found.'));
 	}
 	
 	public function actionUnfollow()
@@ -762,7 +771,7 @@ class StartupController extends Controller
 		}
 		
 		else
-			throw new CHttpException(404, 'Page not found.');	
+			throw new CHttpException(404,UserModule::t('Page not found.'));	
 	}
 	
 	public function actionMultPic($name)
@@ -773,7 +782,7 @@ class StartupController extends Controller
 		
 			$model=$this->loadModel($name);
 			if(!Yii::app()->user->checkAccess('editStartup', array('startup'=>$model)))
-				throw new CHttpException(403,'Você não pode editar esse projeto!');
+				throw new CHttpException(403,UserModule::t('You cannot edit this Startup!'));
 			
 			$img_list=CUploadedFile::getInstancesByName('mult_pic');
 			
@@ -830,7 +839,7 @@ class StartupController extends Controller
 		$model=$this->loadModel($name);
 		
 		if(!Yii::app()->user->checkAccess('editStartup', array('startup'=>$model)))
-            throw new CHttpException(403,'Você não pode editar essa startup!');
+            throw new CHttpException(403,UserModule::t('You cannot edit this Startup!'));
 		
 		$user_startup = new UserStartup;
 		
@@ -920,7 +929,7 @@ class StartupController extends Controller
 	{
 		$model=$this->loadModel($name);
 		if(!Yii::app()->user->checkAccess('editStartup', array('startup'=>$model)))
-            throw new CHttpException(403,'Você não pode editar esse projeto!');
+            throw new CHttpException(403,UserModule::t('You cannot edit this Startup!'));
 		$user_startup=UserStartup::model()->find('user_id=:u_id AND startup_id=:s_id', array(':u_id'=>$id, ':s_id'=>$model->id));
 		$user_startup->delete();
 			
@@ -934,7 +943,7 @@ class StartupController extends Controller
     {
         $startup = Startup::model()->findByPk($sid);
         if(!$startup->hasUserRelation())
-            throw new CHttpException(403,'Você não pode fazer isso.');
+            throw new CHttpException(403,UserModule::t('You cannot edit this Startup!'));
         
         $user_startup=UserStartup::model()->find('user_id=:u_id AND startup_id=:s_id', array(':u_id'=>$uid, ':s_id'=>$sid));
 		$user_startup->approved=1;
@@ -1100,7 +1109,7 @@ class StartupController extends Controller
 			echo json_encode($list);
 		
 		else 
-			throw new CHttpException(403,'Você não pode editar essa startup!');
+			throw new CHttpException(403,UserModule::t('You cannot edit this Startup!'));
 			
 	}
